Move common data structures to common.go
so we can fix the docs
